Upm interfaces2 (#23)
* Added interface installing to cmake.

You can use it by setting IFACE_HDR with the names of the required interfaces or set module_iface in the sensor's CMakeLists.txt file.
Example files: src/abp/CMakeLists.txt src/lm35/CMakeLists.txt


*  Added modifications to cmake to create a separate lib for the new interfaces using SWIG.


* Added new interfaces bindings


* Added bindings for iDistance sensors.


* Added bindings for iDistanceInterrupter sensors.


* Added bindings for iHallEffect sensors.


* Added bindings for iHeartRate sensors.


* dded bindings for iHumidity sensors.


* Added bindings for iLight sensors.
Removed old interfaces C++ examples.


* Added bindings for iMoisture sensors.


* Added bindings for iMotion and iPressure sensors.


* Updated rule for ABP sensor.


* Updated CMake interface examples.


* Updated Java CMake and examples to comply with the new interfaces.


* Removed using namespaces from interface examples


* Modified all iDistance sensors to return float.


* Added documentation to interfaces


* Added getVolts() to iEMG interface and its sensors.


* Added bindings for iElectromagnet interface.


* Removed commented code
Ran clang-format on modified files

+++ include/interfaces/iDistance.hpp
@@ -36,7 +36,13 @@ namespace upm
   {
   public:
     virtual ~iDistance() {}
-    virtual int getDistance() = 0;
+
+    /**
+     * Get distance from sensor
+     * 
+     * @return Returns the distance
+     */
+    virtual float getDistance() = 0;
     /**
       * Convert distance value from Cm(default) to one
       * of the following:
